+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0] - 2026-02-08
+
+### Changed
+
+- Add fullYoastHead field to drive frontend Yoast integration
+- Update plugin version metadata and changelog entries
+
 ## [2.2.4] - 2025-12-03
 
 ### Added